+++ src/lib/UserSetting/Setting/AbstractDateTimeFormat.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
 namespace EzSystems\EzPlatformUser\UserSetting\Setting;
 
 use EzSystems\EzPlatformUser\UserSetting\FormMapperInterface;
-use EzSystems\EzPlatformUser\UserSetting\Tools\DateTimeFormat\FormatterInterface;
+use EzSystems\EzPlatformUser\UserSetting\DateTimeFormat\FormatterInterface;
 use EzSystems\EzPlatformUser\UserSetting\ValueDefinitionInterface;
 use DateTimeImmutable;
 
@@ -18,12 +18,12 @@ abstract class AbstractDateTimeFormat implements ValueDefinitionInterface, FormM
     /** @var \EzSystems\EzPlatformUser\UserSetting\Setting\DateTimeFormatSerializer */
     protected $serializer;
 
-    /** @var \EzSystems\EzPlatformUser\UserSetting\Tools\DateTimeFormat\Formatter|null */
+    /** @var \EzSystems\EzPlatformUser\UserSetting\DateTimeFormat\Formatter|null */
     protected $formatter;
 
     /**
      * @param \EzSystems\EzPlatformUser\UserSetting\Setting\DateTimeFormatSerializer $serializer
-     * @param \EzSystems\EzPlatformUser\UserSetting\Tools\DateTimeFormat\FormatterInterface $formatter
+     * @param \EzSystems\EzPlatformUser\UserSetting\DateTimeFormat\FormatterInterface $formatter
      */
     public function __construct(DateTimeFormatSerializer $serializer, FormatterInterface $formatter)
     {
